addcharacter sql query
dataservice class creates a connection to the sqlite database and contains an addcharacter function that takes in a character object. It will use the connection string for the database and inserts the name, race, class and level entered when creating the character object, into the database. Uses excecute command instead of query command as this is an insert statement. Function returns a value for each row created. If function returns a 1 then you know data was added