editcharacter page prefills data
added the editcharacter page. when you click on the edit button it passed through the ID of that character in that row through to the onget method. The onget method runs the getcharacterbyID function and sets all values of the selected character to a new character object field. The value from this chracter object field are then displayed on the editcharacter view to essentially pre-populate each input tag such as name and race with the current selected character's values.